Moved more admin-related UI files into folder "admin".
