+++ MedicalDocManagment.WebUI/MedicalDocManagment.WebUI.csproj
@@ -201,7 +201,7 @@
   </ItemGroup>
   <ItemGroup>
     <TypeScriptCompile Include="app\admin\admin.component.ts" />
-    <TypeScriptCompile Include="app\models\user-search.model.ts" />
+    <TypeScriptCompile Include="app\admin\models\user-search.model.ts" />
     <TypeScriptCompile Include="app\admin\users-list\page\item-actionlist-delete\confirmation\confirmation.modal.ts" />
     <TypeScriptCompile Include="app\admin\users-list\page\item-actionlist-edit\modal\modal.component.ts" />
     <TypeScriptCompile Include="app\admin\users-list\page\item-actionlist-edit\modal\confirmation\confirmation.modal.ts" />
@@ -257,12 +257,12 @@
   <ItemGroup>
     <TypeScriptCompile Include="app\app.component.ts" />
     <TypeScriptCompile Include="app\app.module.ts" />
-    <TypeScriptCompile Include="app\http.facade.ts" />
+    <TypeScriptCompile Include="app\admin\http.facade.ts" />
     <TypeScriptCompile Include="app\main.ts" />
-    <TypeScriptCompile Include="app\models\positionmodel.ts" />
-    <TypeScriptCompile Include="app\models\rolemodel.ts" />
-    <TypeScriptCompile Include="app\models\usermodel.ts" />
-    <TypeScriptCompile Include="app\models\usersmodel.ts" />
+    <TypeScriptCompile Include="app\admin\models\positionmodel.ts" />
+    <TypeScriptCompile Include="app\admin\models\rolemodel.ts" />
+    <TypeScriptCompile Include="app\admin\models\usermodel.ts" />
+    <TypeScriptCompile Include="app\admin\models\usersmodel.ts" />
     <TypeScriptCompile Include="app\admin\user-add\user.service.ts" />
     <TypeScriptCompile Include="app\admin\user-add\user-add.component.ts" />
     <TypeScriptCompile Include="app\admin\user-add\user.ts" />
