+++ MedicalDocManagment.WebUI/app/admin/user-add/user-add.component.ts
@@ -1,7 +1,7 @@
 ﻿import { Component, OnInit } from '@angular/core';
 import { UserService} from './user.service';
 import {User} from './user';
-import PositionModel from '../../models/positionmodel';
+import PositionModel from '../models/positionmodel';
 import {NotificationsService, SimpleNotificationsComponent} from 'angular2-notifications';
 
 @Component({
